+++ ansi.go
@@ -4,7 +4,11 @@ import (
 	"regexp"
 )
 
-const csi = "\x1b[" // ANSI control sequence introducer
+const (
+	Escape = "\x1b"       // ANSI escape character
+	CSI    = Escape + "[" // ANSI control sequence introducer
+)
+
 var ansiRegex = regexp.MustCompile(`(?:\x1b|\033)\[[;?0-9]*[a-zA-Z]`)
 
 // Strip removes all ANSI escape codes from the input string.