+++ colors.go
@@ -25,56 +25,56 @@ const (
 // Foreground Color
 
 const (
-	Black   = csi + "30m" // Black foreground color.
-	Red     = csi + "31m" // Red foreground color.
-	Green   = csi + "32m" // Green foreground color.
-	Yellow  = csi + "33m" // Yellow foreground color.
-	Blue    = csi + "34m" // Blue foreground color.
-	Magenta = csi + "35m" // Magenta foreground color.
-	Cyan    = csi + "36m" // Cyan foreground color.
-	White   = csi + "37m" // White foreground color.
-	Default = csi + "39m" // Default foreground color.
+	Black   = CSI + "30m" // Black foreground color.
+	Red     = CSI + "31m" // Red foreground color.
+	Green   = CSI + "32m" // Green foreground color.
+	Yellow  = CSI + "33m" // Yellow foreground color.
+	Blue    = CSI + "34m" // Blue foreground color.
+	Magenta = CSI + "35m" // Magenta foreground color.
+	Cyan    = CSI + "36m" // Cyan foreground color.
+	White   = CSI + "37m" // White foreground color.
+	Default = CSI + "39m" // Default foreground color.
 )
 
 // Background Color
 
 const (
-	BlackBg   = csi + "40m" // Black background color.
-	RedBg     = csi + "41m" // Red background color.
-	GreenBg   = csi + "42m" // Green background color.
-	YellowBg  = csi + "43m" // Yellow background color.
-	BlueBg    = csi + "44m" // Blue background color.
-	MagentaBg = csi + "45m" // Magenta background color.
-	CyanBg    = csi + "46m" // Cyan background color.
-	WhiteBg   = csi + "47m" // White background color.
-	DefaultBg = csi + "49m" // Default background color.
+	BlackBg   = CSI + "40m" // Black background color.
+	RedBg     = CSI + "41m" // Red background color.
+	GreenBg   = CSI + "42m" // Green background color.
+	YellowBg  = CSI + "43m" // Yellow background color.
+	BlueBg    = CSI + "44m" // Blue background color.
+	MagentaBg = CSI + "45m" // Magenta background color.
+	CyanBg    = CSI + "46m" // Cyan background color.
+	WhiteBg   = CSI + "47m" // White background color.
+	DefaultBg = CSI + "49m" // Default background color.
 )
 
 // Intense Foreground Color
 
 const (
-	IntenseBlack   = csi + "90m" // Intense Black foreground color.
-	IntenseRed     = csi + "91m" // Intense Red foreground color.
-	IntenseGreen   = csi + "92m" // Intense Green foreground color.
-	IntenseYellow  = csi + "93m" // Intense Yellow foreground color.
-	IntenseBlue    = csi + "94m" // Intense Blue foreground color.
-	IntenseMagenta = csi + "95m" // Intense Magenta foreground color.
-	IntenseCyan    = csi + "96m" // Intense Cyan foreground color.
-	IntenseWhite   = csi + "97m" // Intense White foreground color.
+	IntenseBlack   = CSI + "90m" // Intense Black foreground color.
+	IntenseRed     = CSI + "91m" // Intense Red foreground color.
+	IntenseGreen   = CSI + "92m" // Intense Green foreground color.
+	IntenseYellow  = CSI + "93m" // Intense Yellow foreground color.
+	IntenseBlue    = CSI + "94m" // Intense Blue foreground color.
+	IntenseMagenta = CSI + "95m" // Intense Magenta foreground color.
+	IntenseCyan    = CSI + "96m" // Intense Cyan foreground color.
+	IntenseWhite   = CSI + "97m" // Intense White foreground color.
 
 )
 
 // Intense Background Color
 
 const (
-	IntenseBlackBg   = csi + "100m" // Intense Black background color.
-	IntenseRedBg     = csi + "101m" // Intense Red background color.
-	IntenseGreenBg   = csi + "102m" // Intense Green background color.
-	IntenseYellowBg  = csi + "103m" // Intense Yellow background color.
-	IntenseBlueBg    = csi + "104m" // Intense Blue background color.
-	IntenseMagentaBg = csi + "105m" // Intense Magenta background color.
-	IntenseCyanBg    = csi + "106m" // Intense Cyan background color.
-	IntenseWhiteBg   = csi + "107m" // Intense White background color.
+	IntenseBlackBg   = CSI + "100m" // Intense Black background color.
+	IntenseRedBg     = CSI + "101m" // Intense Red background color.
+	IntenseGreenBg   = CSI + "102m" // Intense Green background color.
+	IntenseYellowBg  = CSI + "103m" // Intense Yellow background color.
+	IntenseBlueBg    = CSI + "104m" // Intense Blue background color.
+	IntenseMagentaBg = CSI + "105m" // Intense Magenta background color.
+	IntenseCyanBg    = CSI + "106m" // Intense Cyan background color.
+	IntenseWhiteBg   = CSI + "107m" // Intense White background color.
 )
 
 // 8-bit foreground color
@@ -84,7 +84,7 @@ func Foreground8Bit(color int) string {
 		return ""
 	}
 
-	return csi + "38;5;" + strconv.Itoa(color) + "m"
+	return CSI + "38;5;" + strconv.Itoa(color) + "m"
 }
 
 // 8-bit background color
@@ -94,7 +94,7 @@ func Background8Bit(color int) string {
 		return ""
 	}
 
-	return csi + "48;5;" + strconv.Itoa(color) + "m"
+	return CSI + "48;5;" + strconv.Itoa(color) + "m"
 }
 
 // 24-bit foreground color
@@ -104,7 +104,7 @@ func Foreground24Bit(r, g, b int) string {
 		return ""
 	}
 
-	return csi + "38;2;" + strconv.Itoa(r) + ";" + strconv.Itoa(g) + ";" + strconv.Itoa(b) + "m"
+	return CSI + "38;2;" + strconv.Itoa(r) + ";" + strconv.Itoa(g) + ";" + strconv.Itoa(b) + "m"
 }
 
 // 24-bit background color
@@ -114,5 +114,5 @@ func Background24Bit(r, g, b int) string {
 		return ""
 	}
 
-	return csi + "48;2;" + strconv.Itoa(r) + ";" + strconv.Itoa(g) + ";" + strconv.Itoa(b) + "m"
+	return CSI + "48;2;" + strconv.Itoa(r) + ";" + strconv.Itoa(g) + ";" + strconv.Itoa(b) + "m"
 }